+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/udf/pythonV1/ArrowUtils.scala
@@ -1,4 +1,4 @@
-package edu.uci.ics.texera.workflow.operators.pythonUDF
+package edu.uci.ics.texera.workflow.operators.udf.pythonV1
 
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.AttributeTypeException
