+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/udf/pythonV1/PythonUDFOpDesc.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.texera.workflow.operators.pythonUDF;
+package edu.uci.ics.texera.workflow.operators.udf.pythonV1;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
@@ -20,8 +20,6 @@ import scala.Function1;
 
 import java.util.List;
 
-import static edu.uci.ics.texera.workflow.operators.pythonUDF.PythonUDFType.SupervisedTraining;
-import static edu.uci.ics.texera.workflow.operators.pythonUDF.PythonUDFType.UnsupervisedTraining;
 import static java.util.Collections.singletonList;
 import static scala.collection.JavaConverters.asScalaBuffer;
 
@@ -112,13 +110,13 @@ public class PythonUDFOpDesc extends OperatorDescriptor {
         }
 
         Schema.Builder outputSchemaBuilder = Schema.newBuilder();
-        if (pythonUDFType == SupervisedTraining) {
+        if (pythonUDFType == PythonUDFType.SupervisedTraining) {
             outputSchemaBuilder.add("class", AttributeType.STRING);
             outputSchemaBuilder.add("precision", AttributeType.STRING);
             outputSchemaBuilder.add("recall", AttributeType.STRING);
             outputSchemaBuilder.add("f1-score", AttributeType.STRING);
             outputSchemaBuilder.add("support", AttributeType.STRING);
-        } else if (pythonUDFType == UnsupervisedTraining) {
+        } else if (pythonUDFType == PythonUDFType.UnsupervisedTraining) {
             outputSchemaBuilder.add("output", AttributeType.STRING);
         } else {
             // for pythonUDFType with map and filter, keep the same schema from input
