+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/udf/pythonV1/PythonUDFOpExec.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.texera.workflow.operators.pythonUDF;
+package edu.uci.ics.texera.workflow.operators.udf.pythonV1;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
@@ -27,9 +27,9 @@ import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
-import static edu.uci.ics.texera.workflow.operators.pythonUDF.PythonUDFOpExec.Channel.FROM_PYTHON;
-import static edu.uci.ics.texera.workflow.operators.pythonUDF.PythonUDFOpExec.Channel.TO_PYTHON;
-import static edu.uci.ics.texera.workflow.operators.pythonUDF.PythonUDFOpExec.MSG.*;
+import static edu.uci.ics.texera.workflow.operators.udf.pythonV1.PythonUDFOpExec.Channel.FROM_PYTHON;
+import static edu.uci.ics.texera.workflow.operators.udf.pythonV1.PythonUDFOpExec.Channel.TO_PYTHON;
+import static edu.uci.ics.texera.workflow.operators.udf.pythonV1.PythonUDFOpExec.MSG.*;
 
 public class PythonUDFOpExec implements OperatorExecutor {
