+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/udf/pythonV2/PythonUDFOpDescV2.java
@@ -1,16 +1,18 @@
-package edu.uci.ics.texera.workflow.operators.udf.python.source;
+package edu.uci.ics.texera.workflow.operators.udf.pythonV2;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.google.common.base.Preconditions;
 import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaTitle;
 import edu.uci.ics.amber.engine.common.IOperatorExecutor;
 import edu.uci.ics.amber.engine.operators.OpExecConfig;
+import edu.uci.ics.texera.workflow.common.metadata.InputPort;
 import edu.uci.ics.texera.workflow.common.metadata.OperatorGroupConstants;
 import edu.uci.ics.texera.workflow.common.metadata.OperatorInfo;
 import edu.uci.ics.texera.workflow.common.metadata.OutputPort;
 import edu.uci.ics.texera.workflow.common.operators.ManyToOneOpExecConfig;
 import edu.uci.ics.texera.workflow.common.operators.OneToOneOpExecConfig;
-import edu.uci.ics.texera.workflow.common.operators.source.SourceOperatorDescriptor;
+import edu.uci.ics.texera.workflow.common.operators.OperatorDescriptor;
 import edu.uci.ics.texera.workflow.common.tuple.schema.Attribute;
 import edu.uci.ics.texera.workflow.common.tuple.schema.OperatorSchemaInfo;
 import edu.uci.ics.texera.workflow.common.tuple.schema.Schema;
@@ -22,7 +24,7 @@ import static java.util.Collections.singletonList;
 import static scala.collection.JavaConverters.asScalaBuffer;
 
 
-public class PythonUDFSourceOpDesc extends SourceOperatorDescriptor {
+public class PythonUDFOpDescV2 extends OperatorDescriptor {
 
     @JsonProperty(required = true)
     @JsonSchemaTitle("Python script")
@@ -34,16 +36,21 @@ public class PythonUDFSourceOpDesc extends SourceOperatorDescriptor {
     @JsonPropertyDescription("Run with multiple workers?")
     public Boolean parallel;
 
+    @JsonProperty(required = true, defaultValue = "true")
+    @JsonSchemaTitle("Retain input columns")
+    @JsonPropertyDescription("Keep the original input columns?")
+    public Boolean retainInputColumns;
+
     @JsonProperty()
-    @JsonSchemaTitle("Columns")
-    @JsonPropertyDescription("The columns of the source")
-    public List<Attribute> columns;
+    @JsonSchemaTitle("Extra output column(s)")
+    @JsonPropertyDescription("Name of the newly added output columns that the UDF will produce, if any")
+    public List<Attribute> outputColumns;
 
 
     @Override
     public OpExecConfig operatorExecutor(OperatorSchemaInfo operatorSchemaInfo) {
         Function1<Object, IOperatorExecutor> exec = (i) ->
-                new PythonUDFSourceOpExec(code);
+                new PythonUDFOpExecV2(code);
         if (parallel) {
             return new OneToOneOpExecConfig(operatorIdentifier(), exec);
         } else {
@@ -55,21 +62,36 @@ public class PythonUDFSourceOpDesc extends SourceOperatorDescriptor {
     @Override
     public OperatorInfo operatorInfo() {
         return new OperatorInfo(
-                "Python UDF Source",
+                "Python UDF V2",
                 "User-defined function operator in Python script",
                 OperatorGroupConstants.UDF_GROUP(),
-                scala.collection.immutable.List.empty(),
+                asScalaBuffer(singletonList(new InputPort("", true))).toList(),
                 asScalaBuffer(singletonList(new OutputPort(""))).toList());
     }
 
-
     @Override
-    public Schema sourceSchema() {
+    public Schema getOutputSchema(Schema[] schemas) {
+        Preconditions.checkArgument(schemas.length == 1);
+        Schema inputSchema = schemas[0];
+
         Schema.Builder outputSchemaBuilder = Schema.newBuilder();
 
+        // keep the same schema from input
+        if (retainInputColumns) {
+            outputSchemaBuilder.add(inputSchema);
+        }
+
+
         // for any pythonUDFType, it can add custom output columns (attributes).
-        if (columns != null) {
-            outputSchemaBuilder.add(columns).build();
+        if (outputColumns != null) {
+            if (retainInputColumns) {
+                // check if columns are duplicated
+                for (Attribute column : outputColumns) {
+                    if (inputSchema.containsAttribute(column.getName()))
+                        throw new RuntimeException("Column name " + column.getName() + " already exists!");
+                }
+            }
+            outputSchemaBuilder.add(outputColumns).build();
         }
         return outputSchemaBuilder.build();
     }
