+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/udf/pythonUDF/ArrowUtilsSpec.scala
@@ -1,8 +1,9 @@
-package edu.uci.ics.texera.workflow.operators.pythonUDF
+package edu.uci.ics.texera.workflow.operators.udf.pythonUDF
 
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.AttributeTypeException
 import edu.uci.ics.texera.workflow.common.tuple.schema.{AttributeType, Schema}
+import edu.uci.ics.texera.workflow.operators.udf.pythonV1.ArrowUtils
 import org.apache.arrow.memory.{BufferAllocator, RootAllocator}
 import org.apache.arrow.vector.VectorSchemaRoot
 import org.apache.arrow.vector.types.pojo.{ArrowType, Field}
